V105

+++ utility/repo.py
@@ -1,4 +1,4 @@
-from .models import Parameter,Picture
+from .models import Parameter,Picture,ClipBoardItem,MyLink
 from utility.constants import *
 from django.db.models import Q
 from authentication.repo import PersonRepo
@@ -8,6 +8,89 @@ from .apps import APP_NAME
 from .log import leolog
 
 
+class ClipBoardItemRepo:
+    
+    def __init__(self,*args, **kwargs):
+        self.app_name=""
+        self.request=None
+        self.user=None
+        if 'app_name' in kwargs:
+            self.app_name=kwargs['app_name']
+        else:
+            self.app_name=None
+        if 'request' in kwargs:
+            self.request=kwargs['request']
+            self.user=self.request.user 
+        self.me_person=PersonRepo(request=self.request).me
+        self.objects=ClipBoardItem.objects.filter(person=self.me_person)
+
+
+    def list(self,*args, **kwargs):
+        return self.objects.all()
+     
+    def add_clipboard_item(self,*args,**kwargs):
+        result=FAILED
+        if self.me_person is None:
+            return FAILED
+        clip_board_item=ClipBoardItem(person_id=self.me_person.id,*args,**kwargs)
+        clip_board_item.save()
+
+        clip_board_item_list=ClipBoardItem.objects.filter(person_id=self.me_person)
+        if len(clip_board_item_list)>CLIPBODRD_MAX_LENGTH :
+            clip_board_item_list.first().delete()
+        result=SUCCEED
+        return result
+
+
+
+
+class MyLinkRepo:
+    
+    def __init__(self,*args, **kwargs):
+        self.app_name=""
+        self.request=None
+        self.user=None
+        if 'app_name' in kwargs:
+            self.app_name=kwargs['app_name']
+        else:
+            self.app_name=None
+        if 'request' in kwargs:
+            self.request=kwargs['request']
+            self.user=self.request.user 
+        self.me_person=PersonRepo(request=self.request).me
+        self.objects=MyLink.objects.filter(person=self.me_person).order_by('priority')
+
+
+    def list(self,*args, **kwargs):
+        return self.objects.all()
+    def delete_my_link(self,my_link_id):
+        result,my_links=FAILED,[]
+        my_link=MyLink.objects.filter(person_id=self.me_person.id).filter(id=my_link_id).first()
+        if my_link is not None:
+            my_link.delete()
+            result=SUCCEED
+            my_links= MyLink.objects.filter(person_id=self.me_person.id).order_by('link__priority')
+        return result,my_links
+    def add_my_link(self,*args,**kwargs):
+        result=FAILED
+        if self.me_person is None:
+            return FAILED
+        from attachments.models import Link
+        my_link=MyLink(person_id=self.me_person.id)
+        if 'title' in kwargs:
+            my_link.title=kwargs['title']
+        if 'url' in kwargs:
+            my_link.url=kwargs['url']
+        if 'priority' in kwargs:
+            my_link.priority=kwargs['priority']
+        my_link.person_id=self.me_person.id
+        my_link.save()
+
+        my_link_list=MyLink.objects.filter(person_id=self.me_person).order_by('link__priority')
+        if len(my_link_list)>MY_LINKS_LENGTH :
+            my_link_list.first().delete()
+        result=SUCCEED
+        return result
 
 
 class PictureRepo:
@@ -63,7 +146,6 @@ class PictureRepo:
     def get(self,*args, **kwargs):
         return self.picture(*args, **kwargs)
 
-
 class ParameterRepo:    
     def __init__(self,request,*args, **kwargs):
         self.request=None 
@@ -83,7 +165,7 @@ class ParameterRepo:
         if 'force' in kwargs and kwargs['force']:
             self.objects=Parameter.objects.all()
     
-    def change_parameter(self,*args, **kwargs):
+    def change_parameter_temp_deleted(self,*args, **kwargs):
         if not self.request.user.has_perm(APP_NAME+'.change_parameter'):
             return None
         parameter_id=kwargs['parameter_id'] if 'parameter_id' in kwargs else None
@@ -114,6 +196,9 @@ class ParameterRepo:
         value=kwargs['value']
         name=kwargs['name']
         app_name=kwargs['app_name']
+        if not self.request.user.has_perm(APP_NAME+'.change_parameter'):
+            message='مجوز دسترسی شما برای این کار کافی نمی باشد. '
+            return FAILED,message,None
         Parameter.objects.filter(app_name=app_name).filter(name=name).delete()
         if value is None:
             value=name
@@ -122,7 +207,8 @@ class ParameterRepo:
         parameter.app_name=app_name
         parameter.origin_value=value
         parameter.save()
-        return parameter
+        message="پارامتر با موفقیت تغییر یافت."
+        return SUCCEED,message,parameter
